requirements.txt
Flask==0.12
peewee==2.8.7
python-slugify==1.2.1
pelican==3.7.1
Markdown==2.6.8
# dep
appdirs==1.4.1
blinker==1.4
click==6.7
docutils==0.13.1
feedgenerator==1.9
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
Unidecode==0.4.20
Werkzeug==0.11.15
